Exit the VM when example canvas is closed
This change makes it possible to access the `JFrame` associated with a
canvas. This in turn allows the `Main` class to configure the frame to
exit the VM when the frame is closed. This is turn means if we `run` the
project from SBT we return to the SBT prompt when we close the window.